4.21*

+++ osa4/blogs/controllers/routes.js
@@ -56,8 +56,21 @@ theRouter.put('/:id', async (request, response) => {
 
 
 theRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+      return response.status(401).json({ error: 'invalid user' })
+  }
 })
 
 module.exports = theRouter
